backend/requirements.txt
langchain_huggingface
langchain_community
langchain_google_genai
chromadb
hdbcli
langchain_google_genai
